requirements.txt
pycrypto==2.6
jake==0.1.7
markdown2==2.3.4
